Fixed potential inconsistency in prizes
Before this fix, if a user successfully deposited money, and before
the deposit_callback a prize updated was triggered, then the prize
would be higher than in reallity. This is because the total tickets
were updated in external, but not in our "records" yet.

While fixing this error, I noticed that now we didn't need a the
"give_tickets" function anymore.